README.md
After measuring the condition of the desired land in different layers of the urban form (land use system, movement and access system, urban landscape, physical and environmental form), we classified the obtained information using SWOT analysis. Then we adjusted the vision and goals of the project based on the green infrastructure approach. We presented several conceptual options, by using the goal achievement matrix and the AHP method, one of the conceptual options was selected as a design option. After preparing the structure map of the design areas, agendas were written for each design area and a three-dimensional master plan was prepared.
In the design of the areas, the existing green spaces were considered and connected to each other and a green network was created.
We used rain gardens in the sidewalks that are designed to collect and manage stormwater.
We designed Infiltration Basins to store rain water beneath a surface that closely mimics its surroundings, like grass or concrete. 
Blue roofs were designed without vegetation for the primary purpose of detaining stormwater. 
We used permeable paving instead of traditional impermeable concrete or asphalt which allow water to seep in between the paving materials and be absorbed into the ground.After measuring the condition of the desired land in different layers of the urban form (land use system, movement and access system, urban landscape, physical and environmental form), we classified the obtained information using SWOT analysis. Then we adjusted the vision and goals of the project based on the green infrastructure approach. We presented several conceptual options, by using the goal achievement matrix and the AHP method, one of the conceptual options was selected as a design option. After preparing the structure map of the design areas, agendas were written for each design area and a three-dimensional master plan was prepared. In the design of the areas, the existing green spaces were considered and connected to each other and a green network was created. We used rain gardens in the sidewalks that are designed to collect and manage stormwater. We designed Infiltration Basins to store rain water beneath a surface that closely mimics its surroundings, like grass or concrete. Blue roofs were designed without vegetation for the primary purpose of detaining stormwater. We used permeable paving instead of traditional impermeable concrete or asphalt which allow water to seep in between the paving materials and be absorbed into the ground.
Skills: SPSS Statistics · V-Ray · SketchUp · AutoCAD · Geographic Information Systems (GIS)